Fix hash check
Previously `strconv.ParseUint` was used to validate if a given full SHA1 hash is valid. Since the nature of SHA1 this hash could easily exceed the range of Uint and therefore is not a good solution to make sure a hash is valid hex notation.

The result was the `'hash' is not a valid SHA-1 hash (SQLSTATE HV000)` error even though the given hash was valid.

Instead, we now use `hex.Decode` which is meant for this kind of tasks and will not fail anymore if a given hash is bigger than the Uint limit.